+++ server/routes/index.js
@@ -1,6 +1,7 @@
 var express = require("express");
 var router = express.Router();
 const services = require("../services/render");
+const controller = require("../controller/controller");
 const mysql = require("mysql");
 const sha256 = require("sha256");
 const url = require("url");
@@ -22,253 +23,25 @@ router.get("/", services.homeRoutes);
 router.get("/sign", services.signRoutes);
 
 router.post("/signIn", (req, res) => {
-  console.log(req.body);
-  req.session.login = true;
+  // console.log("signIn====================");
+  // console.log(req.body);
+  req.session.login = req.body.login;
   req.session.name = req.body.name;
 
   res.redirect("/");
 });
 
 /* register api */
-router.post("/register", function (req, res) {
-  console.log("got /register request!");
-  try {
-    const email = req.body.email || "";
-    const password = req.body.password || "";
-    const name = req.body.name || "";
-    const age = req.body.age || "";
-    console.log(
-      `email : ${email}, password : ${password}, name : ${name}, age : ${age}`
-    );
+router.post("/register", controller.registerControll);
 
-    if (!email.length || !password.length || !name.length || !age.length) {
-      return res.status(400).json({ err: "Incorrect info" });
-    }
+router.post("/login", controller.loginPostControll);
 
-    console.log("done");
-    var sql = `SELECT id FROM users WHERE id = "${email}"`;
-    db.query(sql, function (err, rows, fields) {
-      if (err) {
-        console.log(err);
-        return res.status(400).json({ register: false });
-      } else {
-        console.log(`redundanted email is : ${rows}`);
-        if (rows.length) {
-          return res.status(400).json({ email: "Redundanted id" });
-        } else {
-          console.log("Valid email confirmed!");
-          db.query(
-            `INSERT INTO users (id, name, password, age) VALUES("${email}", "${name}", "${sha256(
-              password
-            )}", ${parseInt(age)})`,
-            function (err, rows, fields) {
-              console.log("inserting data into users database.....");
-              if (err) {
-                console.log(err);
-                return res.status(400).json({ register: false });
-              } else {
-                console.log("SUCCESS!");
-                return res.status(201).json({ register: true });
-              }
-            }
-          );
-        }
-      }
-    });
-  } catch (e) {
-    return res.status(500).json({ err: "Server error" });
-  }
-});
+router.get("/login", controller.loginGetControll);
 
-router.post("/login", function (req, res) {
-  console.log("got request!");
-  console.log(req.body.email);
-  console.log(req.body.password);
-  const sql = `SELECT password, name, uid FROM users WHERE id = "${req.body.email}"`;
-  db.query(sql, (err, rows) => {
-    if (err)
-      return res.status(500).json({
-        login: false,
-        result: "sql error",
-        err: err,
-      });
-    else {
-      try {
-        console.log(rows);
-        if (rows[0].password === sha256(req.body.password)) {
-          let token = jwt.sign({ uid: rows[0].uid }, "ang");
-          return res.status(200).json({
-            login: true,
-            name: rows[0].name,
-            token: token,
-          });
-        } else {
-          return res.status(404).json({
-            result: "invalid id",
-          });
-        }
-      } catch (e) {
-        console.log(e);
-        return res.status(400).json({
-          login: false,
-          result: "invalid password",
-        });
-      }
-    }
-  });
-});
+router.get("/user", controller.userGetControll);
 
-router.get("/login", function (req, res) {
-  console.log("got authorization request!");
-  const decodedToken = jwt.verify(
-    req.headers.authorization.split(" ")[0],
-    "ang"
-  )["uid"];
-  const query = `select * from users where uid = ${decodedToken}`;
+router.post("/qna/question", controller.questionPostControll);
 
-  db.query(query, (err, rows, fields) => {
-    if (err) {
-      console.log(err);
-      return res.status(401).json({
-        login: false,
-      });
-    } else {
-      if (!rows.length) {
-        return res.status(401).json({
-          login: false,
-        });
-      } else {
-        return res.status(200).json({
-          login: true,
-          name: rows[0].name,
-        });
-      }
-    }
-  });
-});
-
-router.get("/user", function (req, res) {
-  console.log("got /user request!");
-  try {
-    var parsedUrl = url.parse(req.url);
-    const email = querystring.parse(parsedUrl.query, "&", "=").email;
-
-    if (email.length == 0) {
-      return res.status(400).json({
-        err: "Invalid id",
-      });
-    }
-
-    const sql = `SELECT * FROM users WHERE id = "${email}"`;
-    db.query(sql, function (err, rows, fields) {
-      if (err) {
-        console.log(err);
-        return res.status(400).json({
-          err: "Invalid id",
-        });
-      } else {
-        try {
-          return res.status(200).json({
-            uid: rows[0].uid,
-            name: rows[0].name,
-            age: rows[0].age,
-          });
-        } catch (e) {
-          console.log(e);
-          return res.status(500).json({
-            err: "Server error",
-          });
-        }
-      }
-    });
-  } catch (e) {
-    console.log(e);
-  }
-});
-
-router.post("/qna/question", function (req, res) {
-  console.log("got /qna/question request!");
-  try {
-    const title = req.body.title || "";
-    const writer_id = req.body.writer_id || "";
-    const category = req.body.category || "";
-    const main_text = req.body.main_text || "";
-
-    if (
-      !title.length ||
-      !writer_id.length ||
-      !category.length ||
-      !main_text.length
-    ) {
-      return res.status(401).json({
-        post: false,
-        err: "Invalid field values",
-      });
-    }
-
-    const query = `INSERT INTO questions (title, writer_id, category, timestamp, main_text) VALUES("${title}", "${writer_id}", "${category}", (now()), "${main_text}")`;
-    db.query(query, function (err, rows, fields) {
-      if (err) {
-        return res.status(500).json({
-          post: false,
-          err: err,
-        });
-      } else {
-        db.query(
-          `select pid from questions order by pid desc limit 1`,
-          function (err, rows, fields) {
-            if (err) {
-              return res.status(500).json({
-                post: false,
-                err: err,
-              });
-            } else {
-              return res.status(201).json({
-                post: true,
-                pid: rows[0].pid,
-              });
-            }
-          }
-        );
-      }
-    });
-  } catch (e) {
-    console.log(e);
-  }
-});
-
-router.get("/qna/list", function (req, res) {
-  console.log("got /qna/list request!");
-  var parsedUrl = url.parse(req.url);
-  const page = querystring.parse(parsedUrl.query, "&", "=").page;
-
-  console.log(page);
-
-  if (!page.length) {
-    return res.status(401).json({
-      result: "",
-    });
-  }
-
-  const query = `SELECT * FROM questions ORDER BY pid DESC LIMIT ${
-    (page - 1) * 25
-  }, 25`;
-  db.query(query, function (err, rows, fields) {
-    if (err) {
-      return res.status(500).json({
-        result: "",
-        err: err,
-      });
-    } else {
-      return res.status(200).json({
-        result: rows,
-      });
-    }
-  });
-});
-
-router.get("/test", function (req, res) {
-  console.log("test");
-});
+router.get("/qna/list", controller.qnaListGetControll);
 
 module.exports = router;
